requirements.txt
dashscope>=1.14.0
python-dotenv>=0.19.0
openai>=1.0.0